requirements.txt
Flask
Flask-RESTful
flask-swagger-ui
pandas
PyYAML
